requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.4.1
    # via django
black==22.1.0
    # via -r requirements/dev.in
cfgv==3.3.1
    # via pre-commit
click==8.0.3
    # via
    #   black
    #   pip-tools
distlib==0.3.4
    # via virtualenv
django==3.2.9
    # via -r requirements/base.in
django-environ==0.8.1
    # via -r requirements/base.in
django-node-assets==0.9.9
    # via -r requirements/base.in
django-widget-tweaks==1.4.9
    # via -r requirements/base.in
filelock==3.4.1
    # via virtualenv
identify==2.4.4
    # via pre-commit
importlib-metadata==4.8.2
    # via
    #   click
    #   pep517
    #   pre-commit
    #   virtualenv
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pillow==8.4.0
    # via -r requirements/base.in
pip-tools==6.4.0
    # via -r requirements/dev.in
platformdirs==2.4.0
    # via
    #   black
    #   virtualenv
pre-commit==2.16.0
    # via -r requirements/dev.in
psycopg2-binary==2.9.2
    # via -r requirements/base.in
pytz==2021.3
    # via django
pyyaml==6.0
    # via pre-commit
six==1.16.0
    # via virtualenv
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pre-commit
tomli==1.2.2
    # via
    #   black
    #   pep517
typed-ast==1.5.2
    # via black
typing-extensions==3.10.0.2
    # via
    #   asgiref
    #   black
    #   importlib-metadata
virtualenv==20.13.2
    # via pre-commit
wheel==0.37.0
    # via pip-tools
zipp==3.6.0
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
